disk: shrink esi to 16 bit, two problems, see long
        There are two problems with using the operand size override for
        si in the CHS disk reader. Specifically they are these:

                1. Addressable memory is limited to 1MiB + 64K - 16,
                   which means the maximum read is theoretically capped
                   at 2175 sectors. That is before one even considers
                   that a around half of low memory is actually system
                   reserved. In reality the maximum number of sectors to
                   read is much lower than this.
                2. pusha/popa doesn't protect the upper half of esi from
                   being clobbered by the BIOS, so a sufficiently
                   obnoxious one could easily clobber it and have the
                   disk reader catastrophically corrupt memory.

        For that purpose it is better to just limit the maximum read
        size, and make the limitation a little bit more obvious by
        shrinking the type of the input, e.g. uint16_t blocks.

+++ boot/stage2/rmode.h
@@ -38,6 +38,6 @@ int32_t  bios_mmap(struct e820_info *);
 void     bios_print(char *, size_t);
 uint32_t bios_disk_geometry(struct disk_info *, uint8_t);
 uint32_t bios_disk_reset(uint8_t);
-int32_t  bios_chs_read(struct disk_info *, char *, size_t, uint32_t, uint8_t);
+int32_t  bios_chs_read(struct disk_info *, char *, uint32_t, uint16_t, uint8_t);
 
 #endif
